http/src/requirements.txt
kafka-python>=0.10.2.0
flask